review: split service spec to fix 750-line ESLint error (ROK-826)
Extract grace-period tests and shared setup into separate files:
- recruitment-reminder.service.spec-helpers.ts (makeEventRow, createRecruitmentReminderTestModule)
- recruitment-reminder.service.grace-period.spec.ts (grace period describe block)
- recruitment-reminder.service.spec.ts refactored to import from spec-helpers

Fixes max-lines error (766 → 641 countable lines). All 95 tests pass.
